Move CExpr glue code into sub-folder
This paves the way to implement more Macro languages
