+++ hs-bindgen/src/HsBindgen/Internal/Macro/CExpr/Parse.hs
@@ -1,4 +1,4 @@
-module HsBindgen.Internal.Macro.Parse (
+module HsBindgen.Internal.Macro.CExpr.Parse (
     parseMacro
   ) where
 
@@ -7,8 +7,8 @@ import C.Expr.Parse qualified as CExpr
 import Clang.CStandard
 import Clang.HighLevel.Types
 
-import HsBindgen.Internal.Macro.CExpr (CExpr)
-import HsBindgen.Internal.Macro.CExpr qualified as Macro
+import HsBindgen.Internal.Macro.CExpr.Type (CExpr)
+import HsBindgen.Internal.Macro.CExpr.Type qualified as CExpr
 import HsBindgen.Macro.Error
 import HsBindgen.Macro.Interface qualified as Macro
 
@@ -18,5 +18,5 @@ parseMacro ::
   -> Either MacroParseError (Macro.Unresolved CExpr)
 parseMacro cStd tokens =
     case CExpr.runParser (CExpr.parseMacro cStd) tokens of
-      Right macro -> Right $ Macro.Unresolved $ Macro.ParsedCExpr macro
+      Right macro -> Right $ Macro.Unresolved $ CExpr.Parsed macro
       Left  err   -> Left  $ MacroParseError err.parseError
