+++ hs-bindgen/src/HsBindgen/Internal/Macro/CExpr/Resolution.hs
@@ -1,4 +1,4 @@
-module HsBindgen.Internal.Macro.Resolution (
+module HsBindgen.Internal.Macro.CExpr.Resolution (
     resolveMacro
   ) where
 
@@ -8,8 +8,8 @@ import C.Expr.Syntax qualified as CExpr
 
 import HsBindgen.Frontend.Analysis
 import HsBindgen.Imports
-import HsBindgen.Internal.Macro.CExpr (CExpr)
-import HsBindgen.Internal.Macro.CExpr qualified as Macro
+import HsBindgen.Internal.Macro.CExpr.Type (CExpr)
+import HsBindgen.Internal.Macro.CExpr.Type qualified as CExpr
 import HsBindgen.IR.C qualified as C
 import HsBindgen.Macro.Error
 import HsBindgen.Macro.Interface qualified as Macro
@@ -18,8 +18,8 @@ resolveMacro ::
      Set C.DeclId
   -> Macro.Unresolved CExpr
   -> Either MacroResolutionError (Macro.Resolved CExpr)
-resolveMacro declIds (Macro.Unresolved (Macro.ParsedCExpr macro)) = do
-    resolvedMacro <- Macro.ParsedCExpr <$> CExpr.annotateMacro resolve macro
+resolveMacro declIds (Macro.Unresolved (CExpr.Parsed macro)) = do
+    resolvedMacro <- CExpr.Parsed <$> CExpr.annotateMacro resolve macro
     pure Macro.Resolved{
         macro = resolvedMacro
       , deps  = getDependencies resolvedMacro
@@ -30,7 +30,7 @@ resolveMacro declIds (Macro.Unresolved (Macro.ParsedCExpr macro)) = do
       CExpr.NameOrdinary nm ->
         resolveBare nm.getIdentifier
       CExpr.NameTagged nm tag ->
-        resolveTagged nm.getIdentifier (Macro.convertTagKind tag)
+        resolveTagged nm.getIdentifier (CExpr.convertTagKind tag)
 
     -- | Resolves a bare name found in a parsed macro body
     --
@@ -87,7 +87,7 @@ resolveMacro declIds (Macro.Unresolved (Macro.ParsedCExpr macro)) = do
         taggedId :: C.DeclId
         taggedId = C.DeclId (C.DeclName nm $ C.NameKindTagged tag) False
 
-getDependencies :: Macro.Parsed CExpr C.DeclId -> [(C.DeclId, Dependency)]
+getDependencies :: CExpr.Parsed CExpr C.DeclId -> [(C.DeclId, Dependency)]
 getDependencies resolvedMacro =
     case resolvedMacro.unwrap of
       CExpr.Macro _ _ _ expr -> goExpr NeedsShape expr
